adding dtype specification to subtract() in torch backend

+++ ivy/functional/backends/torch/elementwise.py
@@ -420,8 +420,8 @@ def subtract(
         x2 = x2.to(promoted_type)
         return torch.subtract(x1, x2, out=out)
     return torch.subtract(
-        x1 if isinstance(x1, torch.Tensor) else torch.tensor(x1),
-        x2 if isinstance(x2, torch.Tensor) else torch.tensor(x2),
+        x1 if isinstance(x1, torch.Tensor) else torch.tensor(x1, dtype=x2.dtype),
+        x2 if isinstance(x2, torch.Tensor) else torch.tensor(x2, dtype=x2.dtype),
     )
